% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blank_rows.R
\name{add_blank_rows}
\alias{add_blank_rows}
\title{add_blank_rows}
\usage{
add_blank_rows(.data, .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{A data.frame or tibble}

\item{.before}{A vector of indices}

\item{.after}{A vector of indices}
}
\description{
add_blank_rows adds rows of NA values at particular specified rows of a data.frame or tibble
One can specify either the row indices before which the rows are to be inserted or the 
row indices after which the blank rows are to be inserted
}
