% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package.R
\name{add_package}
\alias{add_package}
\title{Add a package to the packages.R file}
\usage{
add_package(pkgs, homedir = ".", bioC = FALSE)
}
\arguments{
\item{pkgs}{A character vector of packages}

\item{homedir}{Where are these being included}

\item{bioC}{Whether packages are from Bioconductor}
}
\description{
Packages can be in Github format ('user/packagename') as well
as pure 'packagename'. If bioC=TRUE, all the packages must be from Bioconductor
}
\examples{
add_package(c('tidyverse','broom','stringr'))
}
