% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cols.R
\name{clean_cols}
\alias{clean_cols}
\title{Clean columns to make better tables}
\usage{
clean_cols(x, ..., arrange = F)
}
\arguments{
\item{x}{A data.frame object}

\item{...}{Columns to clean}

\item{arrange}{Whether data should be sorted by column values}
}
\description{
Clean columns to make better tables
}
