% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9.R
\name{expand_code}
\alias{expand_code}
\title{Expanding ICD-9 codes from shortcuts}
\usage{
expand_code(x)
}
\arguments{
\item{x}{ICD-9 codes or ranges (character)}
}
\value{
Expansion of valid ICD-9 codes
}
\description{
Expanding ICD-9 codes from shortcuts
}
