% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{extract_names}
\alias{extract_names}
\title{Extracting package names from strings}
\usage{
extract_names(pkg)
}
\arguments{
\item{pkg}{Input string containing package name}
}
\value{
A string representing the package name
}
\description{
Options exist for Bioconductor and GitHub
}
