% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_onedrive.R
\name{find_onedrive}
\alias{find_onedrive}
\title{Finding the OneDrive directory on your system}
\usage{
find_onedrive(home = NULL)
}
\arguments{
\item{home}{(default=NULL) The home directory under which the OneDrive folder exists. If it is NULL, the system HOME folder is used (`Sys.getenv("HOME")`)}
}
\value{
The normalized path name of the OneDrive folder
}
\description{
If there are more than one folder with "OneDrive" in it, the
user is prompted for a choice
}
