% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd9.R
\name{icd9_codes}
\alias{icd9_codes}
\title{Expand a set of ICD-9 shortcuts into a full set of ICD-9 codes}
\usage{
icd9_codes(x)
}
\arguments{
\item{x}{Set of ICD-9 specifications (character)}
}
\value{
A vector of ICD-9 codes
}
\description{
Expand a set of ICD-9 shortcuts into a full set of ICD-9 codes
}
\examples{
icd9_codes('410-414, V4581')
}
