% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Loading project-required packages in R}
\usage{
load_packages(homedir = ".")
}
\description{
This function reads a file "pkgs.yml" in the `lib` directory of the project
and loads them in R, installing them if necessary
}
\details{
Non-CRAN packages can be accessed via GitHub by specifying `user/package` format
Particular versions of packages can be ensured by specifying `package => version` format
}
