% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myformat_pvalue.R
\name{myformat_pvalue}
\alias{myformat_pvalue}
\title{myformat_pvalue}
\usage{
myformat_pvalue(x, digits = 3)
}
\arguments{
\item{x}{A vector of p-values}

\item{digits}{Number of digits to present in the output}
}
\value{
A character vector of formatted p-values
}
\description{
This function ensures that all the p-values are formatted with the right number of
characters. Below the `digits` threshold, they are presented in scientific notation,
otherwise they are rounded to `digits`
}
