% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{reload}
\alias{reload}
\title{Reloading utility functions into R}
\usage{
reload(d = file.path(getwd(), "lib"))
}
\arguments{
\item{d}{The directory where the utility functions reside. Defaults to the "lib"
sub-directory of the current directory}
}
\value{
Invisibly attaches an environment to the current workspace, named "fn"
}
\description{
This function creates a new environment in R, sources all the R script files
in a directory into that environment, and attaches that environment to the
current namespace. This is a "poor man" packaging system before a formal package is written
}
\details{
This function is meant to load utility functions into R without cluttering the
namespace. It is a "poor man's" packaging withouth formal packaging
}
