% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin_knit.R
\name{spin_knit}
\alias{spin_knit}
\title{spin_knit}
\usage{
spin_knit(f)
}
\arguments{
\item{f}{A R script file, ending in ".R"}
}
\value{
(Invisibly) a HTML document
}
\description{
Start with a R file with "spin" annotation, and derive a rendered document using
rmarkdown::render rather than knit::knit2html
}
