% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_results.coxph.R
\name{table_results.coxph}
\alias{table_results.coxph}
\title{table_results}
\usage{
table_results.coxph(mod, lbls = NULL, tidy = F)
}
\arguments{
\item{mod}{A `coxph` object}

\item{lbls}{An optional named vector providing readable labels for each predictor variable}

\item{tidy}{A boolean (default=F) determining if the output is a tidy tibble or a untidy printable table. Use `tidy = T` if you want to further manipulate the table, or add other results to this table using joins, for example}
}
\value{
A tibble
}
\description{
This version creates a publishable results table from a Cox regression 
run using the `survival::coxph` function. Features include:
}
\details{
+ Adding the reference class for each categorical predictor
+ Choosing to report 95% confidence intervals or p-values (TODO)
+
}
\examples{
library(survival)

}
