% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_packages.R
\name{write_packages}
\alias{write_packages}
\title{iitle}
\usage{
write_packages(pkgs = NULL, homedir = ".")
}
\arguments{
\item{pkgs}{A character array of packages to initialize the project with}

\item{homedir}{The top-level directory of the project, defaults to the current directory}
}
\value{

}
\description{
iitle
}
